Refactor SamlLoginPage into LoginRedirect
